requirements.txt
spacy
newspaper3k
lxml_html_clean